Move query by interface Id test.
