+++ ledger-service/http-json/src/it/scala/http/HttpServiceIntegrationTest.scala
@@ -20,7 +20,6 @@ import org.scalatest.{Assertion, BeforeAndAfterAll}
 import scalaz.\/-
 import shapeless.record.{Record => ShRecord}
 import spray.json.JsValue
-import spray.json._
 
 import scala.concurrent.Future
 
@@ -57,11 +56,6 @@ abstract class HttpServiceIntegrationTest
     discard { dummyFile.delete() }
     discard { staticContentDir.delete() }
     super.afterAll()
-  }
-
-  // nested like this similar to TpId so the references look nicer
-  object CIou {
-    val CIou: domain.TemplateId.OptionalPkg = domain.TemplateId(None, "CIou", "CIou")
   }
 
   "should serve static content from configured directory" in withHttpService {
@@ -268,55 +262,6 @@ abstract class HttpServiceIntegrationTest
           ) =>
         lookup should include regex raw"Cannot resolve Template Key type, given: InterfaceId\([0-9a-f]{64},IIou,IIou\)"
     }
-  }
-
-  "query POST with empty query" - {
-    "with an interface ID" in withHttpService { fixture =>
-      import com.daml.http.json.JsonProtocol._
-      for {
-        _ <- uploadPackage(fixture)(ciouDar)
-        aliceH <- fixture.getUniquePartyAndAuthHeaders("Alice")
-        (alice, aliceHeaders) = aliceH
-        _ <- postCreateCommand(
-          iouCommand(alice, CIou.CIou),
-          fixture,
-          aliceHeaders,
-        )
-
-        searchResp <- search(
-          List.empty,
-          Map(
-            "templateIds" -> Seq(TpId.IIou.IIou).toJson,
-            "query" -> spray.json.JsObject(),
-          ).toJson.asJsObject,
-          fixture,
-          aliceHeaders,
-        )
-      } yield inside(searchResp) {
-        case domain.OkResponse(Seq(ac), None, StatusCodes.OK) => {
-          discard { ac.templateId shouldBe TpId.IIou.IIou.copy(packageId = ac.templateId.packageId) }
-          ac.payload shouldBe spray.json.JsObject()
-        }
-      }
-    }
-  }
-
-
-  private[this] val (_, ciouVA) = {
-    val iouT = ShRecord(issuer = VA.party, owner = VA.party, amount = VA.text)
-    VA.record(Ref.Identifier assertFromString "none:Iou:Iou", iouT)
-  }
-
-  private[this] def iouCommand(party: domain.Party, templateId: domain.TemplateId.OptionalPkg) = {
-    val issuer = Ref.Party assertFromString domain.Party.unwrap(party)
-    val iouT = argToApi(ciouVA)(
-      ShRecord(
-        issuer = issuer,
-        owner = issuer,
-        amount = "42",
-      )
-    )
-    domain.CreateCommand(templateId, iouT, None)
   }
 
   private[this] def iouTransfer[Inj](
