+++ ledger-service/http-json/src/itlib/scala/http/AbstractHttpServiceIntegrationTest.scala
@@ -38,6 +38,7 @@ import spray.json._
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Success
 import com.daml.lf.{value => lfv}
+import com.daml.scalautil.Statement.discard
 import com.google.protobuf.struct.Struct
 import lfv.test.TypedValueGenerators.{ValueAddend => VA}
 
@@ -155,6 +156,11 @@ abstract class AbstractHttpServiceIntegrationTestTokenIndependent
   }
   import HttpServiceTestFixture.{UseTls, accountCreateCommand, archiveCommand}
   import json.JsonProtocol._
+  import AbstractHttpServiceIntegrationTestFuns.ciouDar
+
+  object CIou {
+    val CIou: domain.TemplateId.OptionalPkg = domain.TemplateId(None, "CIou", "CIou")
+  }
 
   override def useTls = UseTls.NoTls
 
@@ -280,6 +286,35 @@ abstract class AbstractHttpServiceIntegrationTestTokenIndependent
           .map(acl => acl.size shouldBe 2)
       } yield {
         assert(true)
+      }
+    }
+
+    "with an interface ID" in withHttpService { fixture =>
+      import com.daml.http.json.JsonProtocol._
+      for {
+        _ <- uploadPackage(fixture)(ciouDar)
+        aliceH <- fixture.getUniquePartyAndAuthHeaders("Alice")
+        (alice, aliceHeaders) = aliceH
+        _ <- postCreateCommand(
+          iouCommand(alice, CIou.CIou),
+          fixture,
+          aliceHeaders,
+        )
+
+        searchResp <- search(
+          List.empty,
+          Map(
+            "templateIds" -> Seq(TpId.IIou.IIou).toJson,
+            "query" -> spray.json.JsObject(),
+          ).toJson.asJsObject,
+          fixture,
+          aliceHeaders,
+        )
+      } yield inside(searchResp) {
+        case domain.OkResponse(Seq(ac), None, StatusCodes.OK) => {
+          discard { ac.templateId shouldBe TpId.IIou.IIou.copy(packageId = ac.templateId.packageId) }
+          ac.payload shouldBe spray.json.JsObject()
+        }
       }
     }
   }
